error correction

+++ Watch.Functions/Entities/TimeEntity.cs
@@ -9,7 +9,7 @@ namespace Watch.Functions.Entities
     {
         public int EmployeeId { get; set; }
         public DateTime Date { get; set; }
-        public byte Type { get; set; }
+        public int Type { get; set; }
         public bool IsConsolidated { get; set; }
     
     }
